Add mobile support for fetching health records
This commit introduces the mobile feature to fetch health records. It includes creation of a new mobile module and the utilization of wasm-bindgen for cross-platform compatibility. Moreover, the SDK version in Cargo.toml and relative paths have been updated to reflect the change. To avoid affecting current server functionality, the inclusion of server-dependent modules are now feature-gated.